refactor(relation): split llm-based-relation-extractor into llm-relation-extractor modules
Issue #351: move provider selection, embedding candidate filter, rate limiter,
JSON/Ollama parsing, and provider-specific extract paths under
services/llm-relation-extractor/. Keep LLMBasedRelationExtractor as orchestrator
with thin private wrappers for existing unit-test spies.

Add design note at docs/superpowers/specs/2026-05-10-issue-351-llm-relation-extractor-design.md.
